requirements.txt
pyyaml~=5.4.1
aiogram~=2.13